# specs/002-phase1-mvp-pipeline/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AX Content Hub API
  description: |
    AX 콘텐츠 큐레이션 봇 API

    ## Internal Endpoints
    `/internal/*` 엔드포인트는 Cloud Run IAM + OIDC 토큰으로 보호됩니다.
    Cloud Scheduler와 Cloud Tasks에서만 호출 가능합니다.

    ## MVP Scope (Phase 1)
    - 소스/구독 관리는 내부용 (인증 없음)
    - 멀티 워크스페이스는 Phase 3에서 구현
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://ax-content-hub-api-xxxxx.run.app
    description: Production (Cloud Run)

paths:
  /health:
    get:
      summary: Health Check
      description: 서버 상태 확인
      operationId: getHealth
      tags:
        - System
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /internal/collect:
    post:
      summary: Trigger Content Collection
      description: |
        Cloud Scheduler에서 1시간마다 호출.
        등록된 모든 활성 소스에서 새 콘텐츠를 수집합니다.
      operationId: triggerCollection
      tags:
        - Internal
      security:
        - oidcToken: []
      responses:
        "200":
          description: Collection triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionResponse"
        "401":
          description: Unauthorized (missing/invalid OIDC token)
        "500":
          description: Internal server error

  /internal/distribute:
    post:
      summary: Trigger Digest Distribution
      description: |
        Cloud Scheduler에서 매일 09:00에 호출.
        발송 예정인 모든 구독에 다이제스트를 발송합니다.
      operationId: triggerDistribution
      tags:
        - Internal
      security:
        - oidcToken: []
      responses:
        "200":
          description: Distribution triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DistributionResponse"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error

  /internal/tasks/process:
    post:
      summary: Process Content Task
      description: |
        Cloud Tasks에서 호출.
        개별 콘텐츠의 번역/요약/스코어링을 수행합니다.
      operationId: processContentTask
      tags:
        - Internal
      security:
        - oidcToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessTaskRequest"
      responses:
        "200":
          description: Processing completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessTaskResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "500":
          description: Processing failed

  /api/sources:
    get:
      summary: List Sources
      description: 등록된 콘텐츠 소스 목록 조회
      operationId: listSources
      tags:
        - Sources
      parameters:
        - name: is_active
          in: query
          schema:
            type: boolean
          description: 활성화 상태로 필터링
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/SourceType"
          description: 소스 타입으로 필터링
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Source"

    post:
      summary: Create Source
      description: 새 콘텐츠 소스 등록
      operationId: createSource
      tags:
        - Sources
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "400":
          description: Validation error

  /api/sources/{source_id}:
    get:
      summary: Get Source
      description: 특정 소스 조회
      operationId: getSource
      tags:
        - Sources
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "404":
          description: Not found

    patch:
      summary: Update Source
      description: 소스 정보 수정
      operationId: updateSource
      tags:
        - Sources
      parameters:
        - $ref: "#/components/parameters/SourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "404":
          description: Not found

    delete:
      summary: Delete Source
      description: 소스 삭제
      operationId: deleteSource
      tags:
        - Sources
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /api/subscriptions:
    get:
      summary: List Subscriptions
      description: 등록된 구독 목록 조회
      operationId: listSubscriptions
      tags:
        - Subscriptions
      parameters:
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"

    post:
      summary: Create Subscription
      description: 새 구독 등록
      operationId: createSubscription
      tags:
        - Subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  /api/subscriptions/{subscription_id}:
    get:
      summary: Get Subscription
      operationId: getSubscription
      tags:
        - Subscriptions
      parameters:
        - $ref: "#/components/parameters/SubscriptionId"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: Not found

    patch:
      summary: Update Subscription
      operationId: updateSubscription
      tags:
        - Subscriptions
      parameters:
        - $ref: "#/components/parameters/SubscriptionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: Not found

    delete:
      summary: Delete Subscription
      operationId: deleteSubscription
      tags:
        - Subscriptions
      parameters:
        - $ref: "#/components/parameters/SubscriptionId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

components:
  securitySchemes:
    oidcToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Google Cloud OIDC token

  parameters:
    SourceId:
      name: source_id
      in: path
      required: true
      schema:
        type: string
      description: Source ID

    SubscriptionId:
      name: subscription_id
      in: path
      required: true
      schema:
        type: string
      description: Subscription ID

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
        timestamp:
          type: string
          format: date-time
      required:
        - status

    SourceType:
      type: string
      enum:
        - rss
        - youtube

    Source:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: "#/components/schemas/SourceType"
        url:
          type: string
          format: uri
        config:
          type: object
        category:
          type: string
          nullable: true
        language:
          type: string
          default: en
        is_active:
          type: boolean
          default: true
        last_fetched_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - url

    SourceCreate:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/SourceType"
        url:
          type: string
          format: uri
        config:
          type: object
        category:
          type: string
        language:
          type: string
          default: en
      required:
        - name
        - type
        - url

    SourceUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        config:
          type: object
        category:
          type: string
        language:
          type: string
        is_active:
          type: boolean

    Subscription:
      type: object
      properties:
        id:
          type: string
        platform:
          type: string
          default: slack
        platform_config:
          type: object
          properties:
            team_id:
              type: string
            channel_id:
              type: string
          required:
            - team_id
            - channel_id
        preferences:
          $ref: "#/components/schemas/SubscriptionPreferences"
        is_active:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - platform_config
        - preferences

    SubscriptionPreferences:
      type: object
      properties:
        frequency:
          type: string
          enum: [realtime, daily, weekly]
          default: daily
        delivery_time:
          type: string
          pattern: "^[0-2][0-9]:[0-5][0-9]$"
          default: "09:00"
        categories:
          type: array
          items:
            type: string
          default: []
        min_relevance:
          type: number
          minimum: 0.0
          maximum: 1.0
          default: 0.3
        language:
          type: string
          default: ko

    SubscriptionCreate:
      type: object
      properties:
        platform_config:
          type: object
          properties:
            team_id:
              type: string
            channel_id:
              type: string
          required:
            - team_id
            - channel_id
        preferences:
          $ref: "#/components/schemas/SubscriptionPreferences"
      required:
        - platform_config

    SubscriptionUpdate:
      type: object
      properties:
        platform_config:
          type: object
        preferences:
          $ref: "#/components/schemas/SubscriptionPreferences"
        is_active:
          type: boolean

    CollectionResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok, partial, error]
        sources_processed:
          type: integer
        contents_collected:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - status
        - sources_processed
        - contents_collected

    DistributionResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok, partial, error]
        subscriptions_processed:
          type: integer
        digests_sent:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - status
        - subscriptions_processed
        - digests_sent

    ProcessTaskRequest:
      type: object
      properties:
        content_id:
          type: string
        retry_count:
          type: integer
          default: 0
      required:
        - content_id

    ProcessTaskResponse:
      type: object
      properties:
        status:
          type: string
          enum: [completed, failed, skipped]
        content_id:
          type: string
        processing_time_ms:
          type: integer
        error:
          type: string
          nullable: true
      required:
        - status
        - content_id

tags:
  - name: System
    description: System health and status
  - name: Internal
    description: Internal endpoints (Cloud Scheduler/Tasks only)
  - name: Sources
    description: Content source management
  - name: Subscriptions
    description: Subscription management